requirements.txt
alembic==1.14.1
aniso8601==8.0.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==7.1.2
distlib==0.3.9
filelock==3.16.1
Flask==1.1.2
Flask-Migrate==4.1.0
Flask-RESTful==0.3.8
flask-restx==1.3.0
Flask-SQLAlchemy==2.4.3
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.8
MarkupSafe==1.1.1
packaging==24.2
pbr==6.1.0
platformdirs==4.3.6
python-dotenv==1.0.1
pytz==2020.1
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
six==1.15.0
SQLAlchemy==1.3.18
stevedore==5.4.0
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
Werkzeug==1.0.1
